+++ etl_plugins/connectors/stream/__init__.py
@@ -1,0 +1,1 @@
+"""Streaming connectors (Kafka, Kinesis, Pulsar, ...)."""